fix: Large Contentful Element for page.tsx
A test with lighthouse showed that potential improvement can be done on
Largest Contentful Paint element in the main page, which is the profile
picture. The profile picture was loaded with Image component from
NextJS. But that component automatically lazy load image by default.
To resolve this, we added priority option in the Image component.